create main_domain_handler
