+++ nerpa.py
@@ -9,8 +9,8 @@ import subprocess
 import nerpa_init
 
 nerpa_init.init()
-import handle_TE
 import handle_rban
+import predictions_preprocessor
 from logger import log
 
 path_to_exec_dir = os.path.dirname(os.path.abspath(__file__)) + "/"
@@ -215,7 +215,7 @@ def run(args):
     if (args.predictions is not None):
         path_predictions = os.path.abspath(copy_prediction_list(args, main_out_dir))
     else:
-        path_predictions = handle_TE.create_predictions_by_antiSAMSHout(get_antismash_list(args), main_out_dir)
+        path_predictions = predictions_preprocessor.create_predictions_by_antiSAMSHout(get_antismash_list(args), main_out_dir)
 
     directory = os.path.dirname(main_out_dir)
     if not os.path.exists(directory):
